requirements.txt
discord.py
cytoolz
